# Improve generator performance