**What this PR does / why we need it**:

The _Determine ARM resource relationships_ stage was taking around a minute to run on my laptop, so I took a look at it - and discovered it's performance was O(n<sup>2</sup>); I modified it to run in O(n) time.

Performance change on my laptop:

| Before  | After  |
| -------:| ------:|
| 1m9.73s | 6.378s |
|  52.48s | 3.014s |

Performance change in CI:

| Before  | After  |
| -------:| ------:|
| 15.258s | 1.333s |
| 15.239s |        |

**Special notes for your reviewer**:

The `findChildren()` method is no longer strictly necessary, but retaining it meant I could keep the tests largely intact.

**How does this PR make you feel**:
![gif](https://media.giphy.com/media/v1.Y2lkPTc5MGI3NjExbnR5MG84bm5yb3UwdW9pamY2ejd4dzRheHRmM3hlY2thdmdzNHJsNyZlcD12MV9naWZzX3NlYXJjaCZjdD1n/ZSvAC5c9RkW4g/giphy.gif)

**If applicable**:
- [x] this PR contains tests